FIXED REMOVAL OF UNSAFE TILES
Best idea I could come up with is it foreach the list of generic unsafe tiles but compare it with the position of the current bomb that has exploded. This lets me find the correct item in the list.
There is a bug where the tiles can detect across the lava and do not delete afterwards, but this is because it should not be going across the lava in the first place.